BigDecimal

+++ src/main/java/com/aventador/domain/CardSlot.java
@@ -3,13 +3,15 @@ package com.aventador.domain;
 import com.aventador.util.GetCash;
 import com.aventador.util.PutCash;
 
+import java.math.BigDecimal;
+
 public final class CardSlot {
 
     //卡槽中刚被放入/刚被吐出的金额数
-    public static double cashAmountJustNow;
+    public static BigDecimal cashAmountJustNow;
 
     //ATM今日可取的总金额数
-    public static double totalCash;
+    public static BigDecimal totalCash;
 
     public CardSlot() {
